segundo Commit

+++ Troca_Base.py
@@ -1,4 +1,5 @@
-#criando classe
+#IMPORT
+import logging
 from msilib.schema import ServiceControl
 from sqlite3 import Cursor
 from sre_parse import State
@@ -9,67 +10,62 @@ from turtle import width
 from firebirdsql import *
 import shutil
 import os as os
-import subprocess
 import sys
 import time
 from setuptools import Command
-import wmi
 import win32service
 import win32serviceutil
 
 #1-ESTRUTURA DA JANELA:
 janela = Tk()
 janela.title("Gdoor - Sistemas")
-#janela.wm_iconbitmap('troca.ico')
 janela["background"]= "orange red"
 
-#0-Dimensões da janela
+#1.1-Dimensões da janela
 largura = 482
 altura = 280
 
-#0.1-RESOLUÇÃO DO NOSSO SISTEMA
+#1.2-RESOLUÇÃO DO NOSSO SISTEMA
 largura_janela = janela.winfo_screenwidth()
 altura_janela = janela.winfo_screenheight()
 
-#0.2- POSIÇÃO DA JANELA
+#1.3- POSIÇÃO DA JANELA
 posx = largura_janela/2 - largura/2
 posy = altura_janela/2 - altura/2
 
-#0.3- DEFINIR A GEOMETRY
+#1.4- DEFINIR A GEOMETRY
 janela.geometry("%dx%d+%d+%d" % (largura,altura,posx,posy))
 
-#0.4- DEFINIR ESPAÇO DO STATUS
+#1.5- DEFINIR ESPAÇO DO STATUS
 LabelFrame  = LabelFrame(janela,text="STATUS:",font=("Arial",16,"bold"), background="orange red", labelanchor=N)
 LabelFrame.pack(fill="both", expand="yes")
 
-#1.1-TEXTO STATUS
+#1.6-TEXTO STATUS
 texto_status = Label(LabelFrame,text="",font=("Arial",12),background="orange red")
 texto_status.pack(anchor=CENTER, fill=Y, expand="yes")
 
-#1.2-TEXTO EXECUÇÃO TROCA DE SENHA 
+#1.7-TEXTO EXECUÇÃO TROCA DE SENHA 
 texto_senha = Label(LabelFrame, text="",font=("Arial",12),background="orange red")
 texto_senha.pack(anchor=CENTER, fill=Y, expand="yes")
 
-#1.3-TEXTO SELEÇÃO COMBOBOX
+#1.8-TEXTO SELEÇÃO COMBOBOX
 seleciona_modulo = Label(janela,text='SELECIONA O MÓDULO GDOOR:',font=("Arial",12,"bold"),background="orange red")
-#seleciona_modulo.place(x=0,y=118)
 seleciona_modulo.pack( expand="yes")
 
-#1.4-VARIAVEL GLOBAL DOS MÓDULOS
+#1.9-VARIAVEL GLOBAL DOS MÓDULOS
 lista_modulos = ["PRO", "SLIM","MEI"]
 caminho_pro='C:/GDOOR Sistemas/GDOOR PRO/'
 caminho_slim='C:/GDOOR Sistemas/GDOOR SLIM/'
 caminho_mei='C:/GDOOR Sistemas/GDOOR MEI/'
 
-#1.5-CRIAÇÃO COMBOBOX
+#1.10-CRIAÇÃO COMBOBOX
 cb_modulos = ttk.Combobox(janela,values=lista_modulos)
 cb_modulos.set(lista_modulos[0])
-#cb_modulos.place(x=260,y=120,width=220)
 cb_modulos.pack(fill="both", expand="yes")
 combo = 0
 
-#2-FUNÇÕES:
-#2.1-VERIFICA SE O DIRETORIO JA ESTA CRIADO, SE NÃO TIVER SERÁ CRIADO
+
+#2-VERIFICA SE O DIRETORIO JA ESTA CRIADO, SE NÃO TIVER SERÁ CRIADO
 if os.path.exists(f'{caminho_pro}'):
     if not os.path.exists(f'{caminho_pro}/Backup Tecnico')and not os.path.exists(f'{caminho_pro}/Base Cliente'):
             os.makedirs(f'{caminho_pro}/Base Cliente')
@@ -83,30 +79,36 @@ if os.path.exists(f'{caminho_mei}'):
             os.makedirs(f'{caminho_mei}/Base Cliente')
             os.makedirs(f'{caminho_mei}/Backup Tecnico')
 
-win32service.SERVICE_CONTROL_CONTINUE
-
 #3-FUNÇÕES
-
-#3.0-PARAR/INICIAR O GUARD E FIREBIRD
-#3.0.1 - PARAR O GUARD
+#3.1 - PARAR O GUARD
 def PARAR_GUARD():
     win32serviceutil.StopService("gdoorGuard_serv")
 def INICIAR_GUARD():
-    time.sleep(1)  
     win32serviceutil.StartService("gdoorGuard_serv")
 
-#3.0.2 - PARARO FIREBIRD    
-
+#3.2 - PARARO/INICIO FIREBIRD    
 def PARAR_FIREBIRD():
     win32serviceutil.StopService("FirebirdGuardianDefaultInstance")
 def INICIAR_FIREBIRD():
-    time.sleep(1)  
     win32serviceutil.RestartService("FirebirdGuardianDefaultInstance") 
-     
-#3.1-CRIAÇÃO DA FUNÇÃO TROCAR
+
+#3.3- VERIFICAÇÃO DOS SERVIÇOS
+def verifica_servicos_guard():
+    verifica_guard= win32serviceutil.QueryServiceStatus('gdoorGuard_serv')[1]
+    if verifica_guard ==1:
+        return False
+    elif verifica_guard ==4:
+        return True
+def verifica_servicos_firebird():
+    verifica_firebird= win32serviceutil.QueryServiceStatus('FirebirdGuardianDefaultInstance')[1]
+    if verifica_firebird ==1:
+        return False
+    if verifica_firebird ==4:
+        return True
+
+#3.4- CRIAÇÃO DA FUNÇÃO TROCAR
 def TROCA():
-    
-    #3.1.1-CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
+    #3.4.0-CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
     combo = cb_modulos.get()
     if combo == lista_modulos[0]:
         modulo = caminho_pro
@@ -114,61 +116,64 @@ def TROCA():
         modulo = caminho_slim
     elif combo == lista_modulos[2]:
         modulo = caminho_mei
+    print(combo)
 
-    #3.1.2- PARAR O GUARD
-    PARAR_GUARD()
-    PARAR_FIREBIRD()
-    time.sleep(2) 
+    #3.4.1- PARAR O GUARD
+    if verifica_servicos_firebird()==True and verifica_servicos_guard()==True:
+        PARAR_GUARD()
+        PARAR_FIREBIRD()
+        time.sleep(5)
+    else:
+        logging.warning('Não parou o serviço no inicio da troca')
 
-    #3.2.4-RENOMEAR O DATAGES
-    os.rename(f'{modulo}/DATAGES.FDB',f'{modulo}/DATAGES_OLD.FDB')
+    #3.4.2- VERIFICA SE OS SERVIÇOS ESTÃO PARADOS A AÇÃO DE TROCA DE BASE
+    if verifica_servicos_firebird()==False and verifica_servicos_guard()==False:
+         #3.4.3- RENOMEAR O DATAGES
+         os.rename(f'{modulo}/DATAGES.FDB',f'{modulo}/DATAGES_OLD.FDB')
+         #3.4.4- BACKUP DO DATAGES DO TECNICO
+         shutil.move(f'{modulo}/DATAGES_OLD.FDB', f'{modulo}/Backup Tecnico/DATAGES_OLD.FDB')
+         #3.4.5- COPIA O DATAGES DA PASTA BASE CLIENTE PARA A PASTA PASTA GDOOR.
+         ori = f'{modulo}Base Cliente/' 
+         des = modulo
+         lista = os.listdir(ori) #lista separando apenas os arquivos do caminho.
+         lista_len = len(lista) #ver quantos arquivos tem na pasta
+         x=0
+         #3.4.6-LAÇO PARA COPIAR OS ARQUIVOS PARA A PASTA DE DESTINO
+         while x < lista_len :
+            caminho_completo_ori = ori + lista[x] 
+            caminho_completo_des = des + lista[x]
+            shutil.move(caminho_completo_ori, caminho_completo_des)
+            x+=1
+         time.sleep(5) 
+    else:
+        logging.warning('Não fez a troca de base')
+    #3.4.3- INICIAR O FIREBIRD
+    if verifica_servicos_firebird()==False:
+        INICIAR_FIREBIRD() 
     
-    #3.3.5-BACKUP DO DATAGES DO TECNICO
-    shutil.move(f'{modulo}/DATAGES_OLD.FDB', f'{modulo}/Backup Tecnico/DATAGES_OLD.FDB')
-      
-    #3.3.6-COPIA O DATAGES DA PASTA BASE CLIENTE PARA A PASTA PASTA GDOOR.
-    ori = f'{modulo}Base Cliente/' 
-    des = modulo
+    #3.4.4- VERIFICA SE O FIREBIRD ESTA INICIADO PARA CONSEGUIR CONECTAR
+    if verifica_servicos_firebird()==True:
+        #3.4.5.1-RODAR O DELETE NA BASE
+        conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
+        cur = conecta.cursor()
+        cur.execute("delete from versao_exe")
+        cur.execute("update usuarios set senha = '1' where usuario = 'ADMINISTRADOR'")
+        conecta.commit() 
+        INICIAR_GUARD()      
+        
+         #3.4.5-APRESENTAR MENSAGEM INDICADO STATUS
+        status = 'Troca de base executada com sucesso!'
+        senha = 'Senha do ADM alterada para 1'
+        texto_status["text"] = status
+        texto_senha["text"] = senha
 
-    lista = os.listdir(ori) #lista separando apenas os arquivos do caminho.
-    lista_len = len(lista) #ver quantos arquivos tem na pasta
-    x=0
-    
-    #3.3.7-LAÇO PARA COPIAR OS ARQUIVOS PARA A PASTA DE DESTINO
-    while x < lista_len:
-        caminho_completo_ori = ori + lista[x] 
-        caminho_completo_des = des + lista[x]
-        shutil.move(caminho_completo_ori, caminho_completo_des)
-        x+=1
-       
-    #3.3.11- INICIAR O GUARD
-    INICIAR_GUARD()
-    INICIAR_FIREBIRD()
-
-    #3.3.8-RODAR O DELETE NA BASE
-    conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
-    cur = conecta.cursor()
-    cur.execute("delete from versao_exe")
-    cur.execute("update usuarios set senha = '1' where usuario = 'ADMINISTRADOR'")
-    conecta.commit()
-   
-    #3.3.9-APRESENTAR MENSAGEM INDICADO STATUS
-    status = 'Troca de base executada com sucesso!'
-    senha = 'Senha do ADM alterada para 1'
-    texto_status["text"] = status
-    texto_senha["text"] = senha
-
-    #3.3.10- DESATIVANDO O BOTÃO CONTRÁRIO
-    TROCA_button.configure(state='disable')
-    REVERTE_button.configure(state='normal')
-    
-    #RODAR O RETAGUARDA NO FINAL
-    #os.system('"%s' % 'gdoorGuard.exe')
-
-#3.2-CRIAÇÃO DA FUNÇÃO REVERTE
+        #3.4.6- DESATIVANDO O BOTÃO CONTRÁRIO
+        TROCA_button.configure(state='disable')
+        REVERTE_button.configure(state='normal')
+        
+#3.5- CRIAÇÃO DA FUNÇÃO REVERTE
 def REVERTER():
-           
-    #3.2.1-CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
+    #3.5.1-CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
     combo = cb_modulos.get()
     if combo == lista_modulos[0]:
         modulo = caminho_pro
@@ -176,54 +181,50 @@ def REVERTER():
         modulo = caminho_slim
     elif combo == lista_modulos[2]:
         modulo = caminho_mei
-
-    #3.2.2- PARAR O GUARD
-    win32service.SERVICE_CONTROL_CONTINUE
-    PARAR_GUARD()
-    PARAR_FIREBIRD()
-    time.sleep(2) 
-
-    #3.2.4-RENOMEAR O DATAGES
-    os.rename(f'{modulo}/DATAGES.FDB',f'{modulo}/DATAGES_CLIENTE.FDB')
     
-    #3.2.5-MOVE O DATAGES DA PASTA GDOOR PARA A PASTA BASE CLIENTE
-    shutil.move(f'{modulo}/DATAGES_CLIENTE.FDB', f'{modulo}/Base Cliente/DATAGES_CLIENTE.FDB')
+    #3.5.2- PARAR O GUARD
+    if verifica_servicos_firebird()==True and verifica_servicos_guard()==True:
+        PARAR_GUARD()
+        PARAR_FIREBIRD()
+        time.sleep(5) 
     
-    #3.2.6-RENOMEAR O DATAGES
-    os.rename(f'{modulo}/Backup Tecnico/DATAGES_OLD.FDB',f'{modulo}/Backup Tecnico/DATAGES.FDB')
+    #3.5.3- VERIFICA OS SERIVOÇOS PARA INICIAR A TROCA DA BASE
+    if verifica_servicos_firebird()==False and verifica_servicos_guard()==False:    
+        #3.3.4-RENOMEAR O DATAGES
+        os.rename(f'{modulo}/DATAGES.FDB',f'{modulo}/DATAGES_CLIENTE.FDB')
+        #3.3.5-MOVE O DATAGES DA PASTA GDOOR PARA A PASTA BASE CLIENTE
+        shutil.move(f'{modulo}/DATAGES_CLIENTE.FDB', f'{modulo}/Base Cliente/DATAGES_CLIENTE.FDB')
+        #3.3.6-RENOMEAR O DATAGES
+        os.rename(f'{modulo}/Backup Tecnico/DATAGES_OLD.FDB',f'{modulo}/Backup Tecnico/DATAGES.FDB')
+        #3.3.7-MOVE O DATAGES DO TECNICO PARA A PASTA DO GDOOR PRO
+        shutil.move(f'{modulo}/Backup Tecnico/DATAGES.FDB', f'{modulo}/DATAGES.FDB')
+        time.sleep(5) 
 
-    #3.2.7-MOVE O DATAGES DO TECNICO PARA A PASTA DO GDOOR PRO
-    shutil.move(f'{modulo}/Backup Tecnico/DATAGES.FDB', f'{modulo}/DATAGES.FDB')
-
-    #3.2.11- INICIAR O GUARD
-    win32service.SERVICE_CONTROL_CONTINUE
-    INICIAR_GUARD()
-    INICIAR_FIREBIRD()
-
-    #3.2.8-RODAR O DELETE NA BASE
-    conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
-    cur = conecta.cursor()
-    cur.execute("delete from versao_exe")
-    cur.execute("update usuarios set senha = '1' where usuario = 'ADMINISTRADOR'")
-    conecta.commit()
-            
-    #3.2.9-APRESENTAR MENSAGEM INDICADO STATUS
-    status = 'Base do tecnico restaurada com sucesso!'
-    senha = 'Senha do ADM alterada para 1'
-    texto_status["text"] = status
-    texto_senha["text"] = senha
-
-    #3.2.10- DESATIVANDO O BOTÃO CONTRÁRIO
-    TROCA_button.configure(state='normal')
-    REVERTE_button.configure(state='disable')
-
-    #RODAR O RETAGUARDA NO FINAL
-    #os.system('"%s' % 'gdoorGuard.exe')
+    #3.5.4- INICIAR O GUARD
+    if verifica_servicos_firebird()==False:
+        INICIAR_FIREBIRD()
     
-#3.3-CRIAÇAO DA FUNÇÃO DELETE_VERSAO
+    #3.5.5-RODAR O DELETE NA BASE
+    if verifica_servicos_guard()==False: 
+        conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
+        cur = conecta.cursor()
+        cur.execute("delete from versao_exe")
+        cur.execute("update usuarios set senha = '1' where usuario = 'ADMINISTRADOR'")
+        INICIAR_GUARD()
+                        
+        #3.5.6- APRESENTAR MENSAGEM INDICADO STATUS
+        status = 'Base do tecnico restaurada com sucesso!'
+        senha = 'Senha do ADM alterada para 1'
+        texto_status["text"] = status
+        texto_senha["text"] = senha
+
+        #3.5.7- DESATIVANDO O BOTÃO CONTRÁRIO
+        TROCA_button.configure(state='normal')
+        REVERTE_button.configure(state='disable')
+                
+#3.6- CRIAÇAO DA FUNÇÃO DELETE_VERSAO
 def DELETE_VERSAO():
-    
-    #3.3.2-CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
+    #3.6.1- CONDICIONAL QUE ARMAZENA O ENDEREÇO NA VARIAVEL modulo
     combo = cb_modulos.get()
     if combo == lista_modulos[0]:
         modulo = caminho_pro
@@ -231,28 +232,30 @@ def DELETE_VERSAO():
         modulo = caminho_slim
     elif combo == lista_modulos[2]:
         modulo = caminho_mei
-    
-    #3.3.1- PARAR O GUARD
-    PARAR_GUARD()
-    time.sleep(2)     
 
-    #3.3.3-RODAR O DELETE NA BASE
-    conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
-    cur = conecta.cursor()
-    cur.execute("delete from versao_exe")
-    conecta.commit()
+    #3.6.2- VERIFICAR OS SERVIÇOS PARA CONSEGUIR CONECTAR COM O BANCO
+    if verifica_servicos_guard() == True:
+        PARAR_GUARD()
+        time.sleep(5)
     
-    #3.3.4-APRESENTAR MENSAGEM INDICADO STATUS
+    #3.6.3- VERIFICAR OS SERVIÇOS E CONECTAR NO BANCO
+    if verifica_servicos_guard() == False:    
+        #3.3.3- RODAR O DELETE NA BASE
+        conecta = firebirdsql.connect(user="SYSDBA", password="masterkey",database=f"{modulo}/DATAGES.FDB",host="localhost")
+        cur = conecta.cursor()
+        cur.execute("delete from versao_exe")
+        conecta.commit()
+        
+    #3.6.4- APRESENTAR MENSAGEM INDICADO STATUS
     status = 'Comando SQL executado com sucesso!'
     texto_status["text"] = status
-
-    #3.3.5-INICIAR O GUARD
-    win32service.SERVICE_CONTROL_CONTINUE
-    INICIAR_GUARD()
-    
-    #RODAR O RETAGUARDA NO FINAL
-    #os.system('"%s' % 'gdoorGuard.exe')
-
+    status_servico = 0
+    time.sleep(5)
+        
+    #3.6.5- INICIAR O GUARD
+    if verifica_servicos_guard()== False:
+        INICIAR_GUARD()
+ 
 #BOTÃO DELETE
 DELETE_button = Button(text='DELETE VERSÃO',height=1,width=15, command=DELETE_VERSAO)
 #DELETE_button.place(x=2,y=150,width=130)
